+++ noodles-cram/CHANGELOG.md
@@ -25,6 +25,8 @@
   * cram/codecs/rans_nx16/encode: Avoid scaling frequencies to 0 during
     normalization.
 
+  * cram/codecs/rans_nx16/encode: Handle 0 RLE symbols.
+
   * cram/codecs/rans_nx16/encode/order_1: Fix order of rANS state
     renormalizations.